Added active stream support

+++ bot.py
@@ -3,9 +3,10 @@ import io
 import os
 import six
 import json
+import queue
+import urllib2
 import argparse
 import threading
-import queue
 import ConfigParser
 from ssl import SSLError
 from google.auth import exceptions
@@ -281,26 +282,31 @@ if __name__ == "__main__":
         args = parser.parse_args()
 
         if args.android:
-            today_folder = os.path.join(os.path.dirname(
-                __file__), 'screenshots', str(strftime("%Y-%m-%d", gmtime())))
-            if not os.path.exists(today_folder):
-                os.makedirs(today_folder)
-            print(
-                info + "Make sure your android device is connected via ADB" + hardreturn)
-            questionNum = 1
-            while questionNum <= 15:
-                screencap_name = strftime(
-                    "%Y-%m-%d_%H:%M:%S", gmtime()) + '_' + str(questionNum) + '.png'
-                raw_input(que + 'Click enter when question is on the screen ')
-                os.system('adb shell screencap -p /sdcard/HQ/' + screencap_name)
-                os.system('adb pull /sdcard/HQ/' + screencap_name)
-                screencap_file = os.path.join(
-                    os.path.dirname(__file__), screencap_name)
-                find_answer(screencap_file)
-                os.rename(screencap_file, os.path.join(
-                    today_folder, screencap_name))
-                os.system('adb shell rm /sdcard/HQ/' + screencap_name)
-                questionNum += 1
+            HQTriviaURL = "https://api-quiz.hype.space/shows/now?type=hq"
+            HQTriviaResponse = json.load(urllib2.urlopen(HQTriviaURL))
+            if HQTriviaResponse["active"]:
+                today_folder = os.path.join(os.path.dirname(
+                    __file__), 'screenshots', str(strftime("%Y-%m-%d", gmtime())))
+                if not os.path.exists(today_folder):
+                    os.makedirs(today_folder)
+                print(
+                    info + "Make sure your android device is connected via ADB" + hardreturn)
+                questionNum = 1
+                while questionNum <= 15:
+                    screencap_name = strftime(
+                        "%Y-%m-%d_%H:%M:%S", gmtime()) + '_' + str(questionNum) + '.png'
+                    raw_input(que + 'Click enter when question is on the screen ')
+                    os.system('adb shell screencap -p /sdcard/HQ/' + screencap_name)
+                    os.system('adb pull /sdcard/HQ/' + screencap_name)
+                    screencap_file = os.path.join(
+                        os.path.dirname(__file__), screencap_name)
+                    find_answer(screencap_file)
+                    os.rename(screencap_file, os.path.join(
+                        today_folder, screencap_name))
+                    os.system('adb shell rm /sdcard/HQ/' + screencap_name)
+                    questionNum += 1
+            else:
+                print(bad + "HQ Trivia is not currently active")
 
         elif args.input_file:
             find_answer(os.path.join(
